Add GA ID to the site
This isn’t probably the best way to do this, but it provides what we
need.

It hard codes site tracking to GA, but we know we’re using GA so I
don’t see that being a problem. It provides flexibility, allowing
user’s to define different UA’s for different machines.